Windows port: normalization: fix narrowing conversions in braced initializers
- groupnorm: explicit static_cast<size_t>/static_cast<uint32_t> for
  int64_t/uint32_t values narrowed into CoreCoord and ttnn::Shape
  braced initializers (groupnorm_grid_utils.cpp, groupnorm_input_mask.cpp)
- layernorm: explicit static_cast<uint32_t> for CoreCoord (size_t) x/y
  fields narrowed into a std::vector<uint32_t> braced return
  (sharded_layernorm_factory_helpers.cpp)
- softmax: explicit static_cast<uint32_t> for size_t block_w/block_h
  products narrowed into DataflowBufferSpec::num_entries
  (softmax_program_factory_attention_optimized_sharded.cpp)

+++ ttnn/cpp/ttnn/operations/normalization/layernorm/device/sharded_layernorm_factory_helpers.cpp
@@ -1584,7 +1584,13 @@ std::vector<uint32_t> reader_sender_named_values(
     } else {
         start_y = core.y - ctx.core_ranges.start_core.y;
     }
-    return {mcast_start.x, mcast_start.y, mcast_end.x, mcast_end.y, start_x, start_y};
+    return {
+        static_cast<uint32_t>(mcast_start.x),
+        static_cast<uint32_t>(mcast_start.y),
+        static_cast<uint32_t>(mcast_end.x),
+        static_cast<uint32_t>(mcast_end.y),
+        start_x,
+        start_y};
 }
 
 // The coordinate block an all-to-all worker walks to reach its remote peers: every X coordinate of
